+++ src/colmap/optim/CMakeLists.txt
@@ -40,6 +40,7 @@ COLMAP_ADD_LIBRARY(
         sparse_cholesky.h sparse_cholesky.cc
         sprt.h sprt.cc
         support_measurement.h support_measurement.cc
+        tiny_solver.h
     PUBLIC_LINK_LIBS
         colmap_math
         colmap_util
@@ -94,4 +95,9 @@ COLMAP_ADD_TEST(
     NAME support_measurement_test
     SRCS support_measurement_test.cc
     LINK_LIBS colmap_optim
+)
+COLMAP_ADD_TEST(
+    NAME tiny_solver_test
+    SRCS tiny_solver_test.cc
+    LINK_LIBS colmap_optim Ceres::ceres Eigen3::Eigen
 )
