+++ src/colmap/optim/tiny_solver.h
