+++ src/colmap/optim/tiny_solver_test.cc
@@ -27,7 +27,7 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include "colmap/math/tiny_solver.h"
+#include "colmap/optim/tiny_solver.h"
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
